Rename index file to js extension
